\name{cccR-package}
\alias{cccR-package}
\alias{cccR}
\docType{package}
\title{
Color contrast calculator with R
}
\description{
A R package focus on color contrast calculator for behavioral ecologist.
}
\details{
\tabular{ll}{
Package: \tab cccR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-08-26\cr
License: \tab GPL 3+\cr
}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
Chen-Pan Liao

Maintainer: Chen-Pan Liao <andrew.43@gmail.com>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
